+++ addons/l10n_hr_fiskal_2_codebooks/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from . import fiskal_document_type
+from . import fiskal_discount_reason
+from . import fiskal_tax_category
+from . import account_tax
+from . import account_move