+++ addons/l10n_hr_kpd/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from . import kpd_code
+from . import product_template
+from . import product_category